@@ -19,20 +19,48 @@ export default class Sample extends Component {
   constructor(props) {
     super(props)
     this.fileIdentifier = `Sample.jsx`
-    this.state = {}
+    this.state = {
+      showTextareaForm: false,
+    }
   }
 
-  componentWillMount = () => tryCatch(this, arguments, () => {})
-  componentDidMount = () => tryCatch(this, arguments, () => {})
-  componentWillReceiveProps = nextProps => tryCatch(this, arguments, () => {})
+  // componentWillMount = () => tryCatch(this, arguments, () => {})
+  // componentDidMount = () => tryCatch(this, arguments, () => {})
+  // componentWillReceiveProps = nextProps => tryCatch(this, arguments, () => {})
   shouldComponentUpdate = (nextProps, nextState) =>
-    !_isEqual(this.props, nextProps) || _isEqual(this.state, nextState)
-  componentWillUpdate = (nextProps, nextState) => tryCatch(this, arguments, () => {})
-  componentDidUpdate = (prevProps, prevState) => tryCatch(this, arguments, () => {})
-  componentWillUnmount = () => tryCatch(this, arguments, () => {})
+    !_isEqual(this.props, nextProps) || !_isEqual(this.state, nextState)
+
+  // componentWillUpdate = (nextProps, nextState) => tryCatch(this, arguments, () => {})
+  // componentDidUpdate = (prevProps, prevState) => tryCatch(this, arguments, () => {})
+  // componentWillUnmount = () => tryCatch(this, arguments, () => {})
+
+  nextButtonOnClick = () => {
+    console.log(`clicked button`)
+    this.setState({
+      showTextareaForm: true,
+    })
+  }
 
   render = () => tryCatch(this, arguments, () => (
     <div className="Sample">
+      <div>{`Please click button continue`}</div>
+      <button
+        className={[`btn`, `primary`].join(` `)} onClick={() => {
+          this.nextButtonOnClick()
+        }}
+      >Next</button>
+
+      <div
+        className={[
+          this.state.showTextareaForm
+          ? null
+          : `hidden`,
+        ].join(` `)}
+      >
+        <h3>{`Please enter random text`}</h3>
+        <textarea rows="10" cols="30" />
+      </div>
+
       <p>{`Your current environment is: ${ENV}`}</p>
     </div>
     ))
