@@ -25,7 +25,7 @@ export default class Example extends Component {
   componentDidMount = () => tryCatch(this, arguments, () => {})
   componentWillReceiveProps = nextProps => tryCatch(this, arguments, () => {})
   shouldComponentUpdate = (nextProps, nextState) =>
-    !_isEqual(this.props, nextProps) || _isEqual(this.state, nextState)
+    !_isEqual(this.props, nextProps) || !_isEqual(this.state, nextState)
   componentWillUpdate = (nextProps, nextState) => tryCatch(this, arguments, () => {})
   componentDidUpdate = (prevProps, prevState) => tryCatch(this, arguments, () => {})
   componentWillUnmount = () => tryCatch(this, arguments, () => {})
